// _app/root/src/modules/interfaces/state/render/IRenderFragment.ts
import IDisplayChart from "../display/IDisplayChart";
import IDisplaySection from "../display/IDisplaySection";
import IRenderFragmentUI from "../ui/IRenderFragmentUI";


export default interface IRenderFragment {

    id: string;
    iKey: string | null;
    iExitKey: string | null;
    exitKey: string | null;
    topLevelMapKey: string;
    mapKeyChain: string;
    guideID: string;
    guidePath: string;
    parentFragmentID: string | null;
    value: string;
    selected: IRenderFragment | null;
    isLeaf: boolean;
    options: Array<IRenderFragment>;

    option: string;
    isAncillary: boolean;
    order: number;

    link: IDisplayChart | null;
    section: IDisplaySection;
    segmentIndex: number | null;

    ui: IRenderFragmentUI;
}